requirements.txt
beautifulsoup4==4.12.2
folium==0.14.0
mysql-connector-python==8.0.29
openai==0.27.4
python-dotenv==1.0.0
requests==2.28.1
streamlit==1.21.0
streamlit_chat==0.0.2.2
streamlit_folium==0.12.0

# -- removed --
# mysql_connector_repackaged==0.3.1
